Move scaling duration computation to Cluster
