+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/applications/Cluster.java
@@ -7,6 +7,7 @@ import com.yahoo.config.provision.ClusterSpec;
 import com.yahoo.vespa.hosted.provision.autoscale.Autoscaler;
 import com.yahoo.vespa.hosted.provision.autoscale.Autoscaling;
 
+import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
@@ -159,6 +160,44 @@ public class Cluster {
     public static Cluster create(ClusterSpec.Id id, boolean exclusive, Capacity requested) {
         return new Cluster(id, exclusive, requested.minResources(), requested.maxResources(), requested.isRequired(),
                            Autoscaling.empty(), Autoscaling.empty(), List.of());
+    }
+
+    /** The predicted time it will take to rescale this cluster. */
+    public Duration scalingDuration(ClusterSpec clusterSpec) {
+        int completedEventCount = 0;
+        Duration totalDuration = Duration.ZERO;
+        for (ScalingEvent event : scalingEvents()) {
+            if (event.duration().isEmpty()) continue;
+            completedEventCount++;
+            // Assume we have missed timely recording completion if it is longer than 4 days
+            totalDuration = totalDuration.plus(maximum(Duration.ofDays(4), event.duration().get()));
+        }
+        if (completedEventCount == 0) { // Use defaults
+            if (clusterSpec.isStateful()) return Duration.ofHours(12);
+            return Duration.ofMinutes(10);
+        }
+        else {
+            Duration predictedDuration = totalDuration.dividedBy(completedEventCount);
+
+            if ( clusterSpec.isStateful() ) // TODO: Remove when we have reliable completion for content clusters
+                predictedDuration = minimum(Duration.ofHours(12), predictedDuration);
+
+            predictedDuration = minimum(Duration.ofMinutes(5), predictedDuration);
+
+            return predictedDuration;
+        }
+    }
+
+    private static Duration minimum(Duration smallestAllowed, Duration duration) {
+        if (duration.minus(smallestAllowed).isNegative())
+            return smallestAllowed;
+        return duration;
+    }
+
+    private static Duration maximum(Duration largestAllowed, Duration duration) {
+        if ( ! duration.minus(largestAllowed).isNegative())
+            return largestAllowed;
+        return duration;
     }
 
 }
